# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Catalog of Products. For this sample, you can use the api key `special-key` to test the authorization     filters.",
        "version": "1.0.0",
        "title": "Catalog of Products",
        "contact": {
            "email": "georgellanque@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://choosealicense.com/licenses/mit/"
        }
    },
    "host": "heroku.catalogofproducts.com",
    "basePath": "/v1.0.0",
    "tags": [{
        "name": "user",
        "description": "Operations about user"
    }, {
        "name": "inventory",
        "description": "Operations about inventory of items"
    }, {
        "name": "products",
        "description": "Operations about products"
    }],
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
        "/user/register": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Register user",
                "description": "This can only be done when the guest wants to be logged.",
                "operationId": "RegisterUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Registered user object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/User"
                    }
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Logs user into the system",
                "description": "",
                "operationId": "loginUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "name": "username",
                        "in": "query",
                        "description": "The user name for login",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password for login in clear text",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "type": "integer",
                                "format": "int32",
                                "description": "calls per hour allowed by the user"
                            },
                            "X-Expires-After": {
                                "type": "string",
                                "format": "date-time",
                                "description": "date in UTC when token expires"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/user/logout": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Logs out current logged in user session",
                "description": "",
                "operationId": "logoutUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/user/{userId}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get user by id",
                "description": "",
                "operationId": "getUserByName",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "username",
                    "in": "path",
                    "description": "The name that needs to be fetched. Use user1 for testing. ",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Invalid username supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "Update user by id",
                "description": "This can only be done by the logged in user.",
                "operationId": "updateUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "name": "username",
                        "in": "path",
                        "description": "name that need to be updated",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Updated user object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid user supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "user"
                ],
                "summary": "Delete user by id",
                "description": "This can only be done by the logged in user.",
                "operationId": "deleteUser",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [{
                    "name": "username",
                    "in": "path",
                    "description": "The name that needs to be deleted",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "400": {
                        "description": "Invalid username supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/inventory/list": {
            "get": {
                "tags": [
                    "inventory"
                ],
                "summary": "Get a list of inventory",
                "description": "",
                "operationId": "listInventory",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/inventory": {
            "post": {
                "tags": [
                    "inventory"
                ],
                "summary": "Register a new item",
                "description": "The user needs a special permission for this action.",
                "operationId": "CreateItem",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Registered item object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/Inventory"
                    }
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/inventory/{itemId}": {
            "put": {
                "tags": [
                    "inventory"
                ],
                "summary": "Updated an existing item",
                "description": "",
                "operationId": "updateItem",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Updated item object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/Inventory"
                    }
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "inventory"
                ],
                "summary": "Delete an existing item",
                "description": "",
                "operationId": "deleteItem",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Delete item object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/Inventory"
                    }
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Get a list of products",
                "description": "",
                "operationId": "listProducts",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Save an item of inventory as a new product",
                "description": "",
                "operationId": "saveProduct",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Save an item object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/Products"
                    }
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/products/{productId}/updateImage": {
            "put": {
                "tags": [
                    "products"
                ],
                "summary": "Update image",
                "description": "",
                "operationId": "updateImageProduct",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "upload an image object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/Products"
                    }
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/products/{productId}": {
            "delete": {
                "tags": [
                    "products"
                ],
                "summary": "Delete an existing product",
                "description": "",
                "operationId": "deleteProduct",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Delete a product object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/Products"
                    }
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/rateproduct/like": {
            "post": {
                "tags": [
                    "RateProduct"
                ],
                "summary": "rate a product as a like",
                "description": "",
                "operationId": "likeProduct",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Like an item object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/RateProduct"
                    }
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/rateproduct/unlike": {
            "post": {
                "tags": [
                    "RateProduct"
                ],
                "summary": "rate a product as an unlike",
                "description": "",
                "operationId": "unlikeProduct",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Unlike an item object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/RateProduct"
                    }
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/rateproduct/unrate": {
            "delete": {
                "tags": [
                    "RateProduct"
                ],
                "summary": "Remove a rate of product",
                "description": "",
                "operationId": "unRateProduct",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Unrate an item object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/RateProduct"
                    }
                }],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "definitions": {
        "User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "description": "User Status"
                }
            }
        },
        "Inventory": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
        "Products": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "idInventoryItems": {
                    "type": "string"
                },
                "imagenUrl": {
                    "type": "string"
                },
                "like": {
                    "type": "integer"
                },
                "unlike": {
                    "type": "integer"
                }
            }
        },
        "RateProduct": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "productId": {
                    "type": "string"
                },
                "admire": {
                    "type": "tinyinteger"
                }
            }
        }
    }
}